pyproject.toml
[tool.poetry]
name = "fastapi-firebase-auth"
version = "0.1.0"
description = ""
authors = ["Joje1283 <joje1283@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
firebase-admin = "^6.4.0"
black = "^24.2.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
loguru = "^0.7.2"
aiohttp = "^3.9.3"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
